requirements.txt
requests
fastapi
uvicorn[standard]
aiohttp
gunicorn
python-lorem
pydantic